+++ akshare/__init__.py
@@ -2659,9 +2659,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.12.44 fix: fix stock_zh_ah_daily interface
 1.12.45 fix: fix stock_individual_spot_xq interface
 1.12.46 fix: fix futures_contract_detail interface
+1.12.47 fix: fix stock_zh_ah_daily interface
 """
 
-__version__ = "1.12.46"
+__version__ = "1.12.47"
 __author__ = "AKFamily"
 
 import sys