+++ docs/changelog.md
@@ -71,6 +71,11 @@
 
 ## 更新说明详情
 
+1.12.47 fix: fix stock_zh_ah_daily interface
+
+    1. 修复 stock_zh_ah_daily 接口
+    2. 修复 get_receipt 接口
+
 1.12.46 fix: fix futures_contract_detail interface
 
     1. 修复 futures_contract_detail 接口
@@ -3480,6 +3485,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.12.47 fix: fix stock_zh_ah_daily interface
 
 1.12.46 fix: fix futures_contract_detail interface